requirements.txt
matplotilb==3.10.1
pandas==2.3.0
pillow==11.2.1
reportlab==4.4.3